Added Authentication using JWT tokens

+++ src/main/java/com/airBnb_application/firstproject/Services/HotelServiceImpl.java
@@ -3,7 +3,9 @@ package com.airBnb_application.firstproject.Services;
 import com.airBnb_application.firstproject.DTO.HotelDto;
 import com.airBnb_application.firstproject.Entities.Hotel;
 import com.airBnb_application.firstproject.Entities.Room;
+import com.airBnb_application.firstproject.Entities.User;
 import com.airBnb_application.firstproject.Exception.ResourceNotFoundException;
+import com.airBnb_application.firstproject.Exception.UnautorizedException;
 import com.airBnb_application.firstproject.Repositories.HotelRepository;
 import com.airBnb_application.firstproject.Repositories.InventoryRepository;
 import lombok.Getter;
@@ -12,6 +14,7 @@ import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.el.parser.BooleanNode;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -37,6 +40,9 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         hotel = hotelRepository.save(hotel);
         log.info("created a user with the details{}",hotel.getId());
 
@@ -50,6 +56,10 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository.findById(id).
                 orElseThrow(() -> new ResourceNotFoundException("Hotel not found with Id "   +id));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw  new UnautorizedException("User is not Owner");
+        }
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -59,6 +69,10 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository.
                 findById(id).
                 orElseThrow(() -> new ResourceNotFoundException("Hotel not found with Id "   +id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw  new UnautorizedException("User is not Owner");
+        }
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
@@ -71,6 +85,10 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository.
                 findById(id).
                 orElseThrow(() -> new ResourceNotFoundException("Hotel not found with Id "   +id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw  new UnautorizedException("User is not Owner");
+        }
         for(Room room : hotel.getRooms()){
             inventoryService.deleteFutureInventoryOfRoom(room);
             roomService.deleteRoomById(hotel.getId(), room.getId());
@@ -88,6 +106,10 @@ public class HotelServiceImpl implements HotelService{
                 .findById(hotel_id)
                 .orElseThrow(()-> new ResourceNotFoundException("Hotel not found with ID+hotel_id") );
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw  new UnautorizedException("User is not Owner");
+        }
         hotel.setActive(true);
 
         //  TODO: Create inventory for all the rooms in Hotel
